Changes to server-side functionality and code organization.
UI and server script are now [fairly] de-coupled. Using JSON for transmission of datasets.
Table creation stored in script, and database is now used effectively as a cache for storing previous operations. If the same gene name (and exact same code for now...) is queried, data from DB is retrieved instead of calculated. Note that later on support for searching should be added so as to facilitate easier lookup and remove the need to enter the promotor/coding region. Searching functionality shouldn't be hard to add. User interface reflects changes to server side transmission of data.